move nav btn to a component for reuse
