+++ src/frontend/src/components/public/PIANavButton/interface.ts
@@ -1,0 +1,5 @@
+import { IPiaForm } from '../../../types/interfaces/pia-form.interface';
+
+export interface INavButton {
+  pia: IPiaForm;
+}
